Implement OpenTelemetry infrastructure
This change begins with the migration for the new OpenTelemetry
infrastructure. The custom metric system is being removed in favor of a
more "handcrafted" observability implementation. This initial change
implements the tracing/otel interop and adds the various layers that
we'll populate in future commits. It'll take some adjustments until we
can leverage the full potential, but this should give us a decent
baseline for future improvements and as starting point for moving all
other telemetry over to the new system.

- Use OptenTelemetry for Telemetry
- Remove OriginContext and OriginEventContext in favor of the
  standardized W3 Baggage and TraceContext APIs
- Remove RootMetricsComponent
- Reinstrument main
- Reinstrument schedulers
- Reinstrument workers