Introduce generic metrics for caches
The naming scheme tries to follows the OpenTelemetry semantic
conventions closely, but we might need to adjust it slightly in the
future. Potentially expensive attribute allocation is zero-cost at
runtime and despite the fairly heavy duration measurments there doesn't
seem to be a relevant performance impact as all operations are in the
nanosecond range when working with a memory-backed `EvictingMap`.

As an example we reinstrument the `EvictingMap` and slightly change its
implementation to more clearly reflect the metrics we care about.