Reimplement metrics for EvictingMap
This introduces a meter for the EvictingMap with various new metrics.
The timelines are labeled with attributes to distinguish between
components. This implementation doesn't yet take the store names from
the config into account. This means that at the moment we have no way to
filter between e.g. AC and CAS stores.